+++ interfaces/HO-Portal/src/app/services/api.service.ts
@@ -165,6 +165,49 @@ export class ApiService {
     );
   }
 
+  patch(
+    endpoint: string,
+    path: string,
+    body: object,
+    anonymous: boolean = false,
+    responseAsBlob: boolean = false,
+    isUpload: boolean = false,
+  ): Promise<any> {
+    const security = this.showSecurity(anonymous);
+    console.log(`ApiService PATCH: ${security} ${endpoint}${path}`, body);
+
+    return new Promise((resolve, reject) =>
+      this.http
+        .patch(endpoint + path, body, {
+          headers: this.createHeaders(isUpload),
+          responseType: responseAsBlob ? 'blob' : null,
+          withCredentials: true,
+        })
+        .pipe(
+          tap((response) =>
+            console.log(
+              `ApiService PATCH: ${security} ${endpoint}${path}:`,
+              body,
+              '\nResponse:',
+              response,
+            ),
+          ),
+          catchError(
+            (error: HttpErrorResponse): Observable<any> =>
+              this.handleError(error, anonymous),
+          ),
+        )
+        .toPromise()
+        .then((response) => {
+          if (response && response.error) {
+            throw response;
+          }
+          return resolve(response);
+        })
+        .catch((err) => reject(err)),
+    );
+  }
+
   delete(
     endpoint: string,
     path: string,