+++ services/121-service/src/programs/programs.controller.ts
@@ -48,6 +48,8 @@ export class ProgramController {
 
   @ApiOperation({ summary: 'Get program by id' })
   @ApiParam({ name: 'programId', required: true, type: 'integer' })
+  // TODO: REFACTOR: Can we make the GET response structure identical to POST body structure by default? Then this setting is not needed anymore.
+  // TODO: REFACTOR: GET /api/programs/:programid with a response body that does not need authorization (i.e. without assigned aid workers) and GET /api/programs/:programid/assigned-aid-workers that requires authorization, see: https://stackoverflow.com/questions/51383267/rest-get-endpoints-returning-different-models-based-on-user-role
   @ApiQuery({
     name: 'formatCreateProgramDto',
     required: false,
@@ -74,6 +76,7 @@ export class ProgramController {
 
   @ApiOperation({ summary: 'Get published programs' })
   @ApiResponse({ status: 200, description: 'Return all published programs.' })
+  // TODO: REFACTOR: into GET /api/programs?published=true
   @Get('published/all')
   public async getPublishedPrograms(): Promise<ProgramsRO> {
     return await this.programService.getPublishedPrograms();
@@ -88,6 +91,7 @@ export class ProgramController {
     status: 401,
     description: 'No user detectable from cookie or no cookie present',
   })
+  // TODO: REFACTOR: into GET /api/users/:userid/programs
   @Get('assigned/all')
   public async getAssignedPrograms(
     @User('id') userId: number,
@@ -143,6 +147,7 @@ export class ProgramController {
   @Permissions(PermissionEnum.ProgramPhaseUPDATE)
   @ApiResponse({ status: 403, description: 'Forbidden.' })
   @ApiParam({ name: 'programId', required: true, type: 'integer' })
+  // TODO: REFACTOR: into PATCH /api/programs/:programid
   @Post(':programId/change-phase')
   public async changePhase(
     @Param() params,
@@ -157,6 +162,7 @@ export class ProgramController {
   @Permissions(PermissionEnum.ProgramUPDATE)
   @ApiOperation({ summary: 'Update program' })
   @ApiParam({ name: 'programId', required: true, type: 'integer' })
+  // TODO: REFACTOR: into PATCH /api/programs/:programid
   @Post(':programId/update')
   public async updateProgram(
     @Param() params,
@@ -171,6 +177,7 @@ export class ProgramController {
   @Permissions(PermissionEnum.ProgramQuestionUPDATE)
   @ApiOperation({ summary: 'Update program question' })
   @ApiParam({ name: 'programId', required: true, type: 'integer' })
+  // TODO: REFACTOR: into PATCH /api/programs/:programid/program-questions/:programquestionid,
   @Post(':programId/update/program-question')
   public async updateProgramQuestion(
     @Body() updateProgramQuestionDto: UpdateProgramQuestionDto,
@@ -203,6 +210,7 @@ export class ProgramController {
   @Permissions(PermissionEnum.ProgramCustomAttributeUPDATE)
   @ApiOperation({ summary: 'Update program custom attributes' })
   @ApiParam({ name: 'programId', required: true, type: 'integer' })
+  // TODO: REFACTOR: into PUT /api/programs/:programid/program-custom-attributes
   @Post(':programId/update/program-custom-attributes')
   public async updateProgramCustomAttributes(
     @Param() params,
@@ -221,6 +229,7 @@ export class ProgramController {
     status: 200,
     description: 'Return PA-table attributes by program-id.',
   })
+  // TODO: REFACTOR: give PA Table attribute data of all phases in the GET /programs/:programid response.
   @Get(':programId/pa-table-attributes/:phase')
   public async getPaTableAttributes(@Param() params): Promise<Attribute[]> {
     return await this.programService.getPaTableAttributes(